Revert "Fix stake rate"
This reverts commit 87ca97bbb0aa275a2cbff05e92ab2136a3799a07.

+++ src/main.h
@@ -34,7 +34,7 @@ static const unsigned int MAX_ORPHAN_TRANSACTIONS = MAX_BLOCK_SIZE/100;
 static const unsigned int MAX_INV_SZ = 500000;
 static const int64_t MAX_MONEY = 10000000000 * COIN;
 
-static const int FORK_HEIGHT = 23200;
+static const int64_t FORK_HEIGHT = 23200;
 
 inline bool MoneyRange(int64_t nValue) { return (nValue >= 0 && nValue <= MAX_MONEY); }
 inline int64_t GetTxFee() { return 0; }
